fix: test application.yml 추가 및 WebMvcConfigurer 통합
