--- backend/src/main/java/com/ody/route/config/ApiCallWebConfig.java
@@ -1,20 +1,0 @@
-package com.ody.route.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class ApiCallWebConfig implements WebMvcConfigurer {
-
-    @Value("${allowed-origins.api-call}")
-    private String[] allowedOrigins;
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/admin/api-call/**")
-                .allowedOrigins(allowedOrigins)
-                .allowedMethods("GET", "POST");
-    }
-}
